usda-ai-dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
databases==0.9.0
jinja2==3.1.4
